Created a function to get the database secret from SSM stored on AWS
