+++ amplify/backend/function/expressServer/src/package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "license": "Apache-2.0",
   "dependencies": {
+    "aws-sdk": "^2.1247.0",
     "aws-serverless-express": "^3.3.5",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.17.1",
